fix(schedule): dedup end-of-day/speaker/duration helpers + desktop day-label TZ bug
- HeaderSection desktop day buttons formatted YYYY-MM-DD via raw new Date().toLocaleDateString, which is parsed as UTC midnight and can render the previous day in western timezones. Now uses the shared timezone-safe formatConferenceDate helper (AGENTS.md forbids raw new Date() for display).
- Hoisted the end-of-day guard into src/lib/schedule/time.ts as withinScheduleEnd; operations.ts and MobileScheduleView.tsx now import it instead of each defining a local copy.
- Extracted populatedSpeakerNames into src/lib/speaker/formatSpeakerNames.ts; MobileScheduleView and DraggableProposal now share it instead of each filtering populated speakers inline.
- Hoisted SERVICE_DURATION_OPTIONS into src/lib/schedule/constants.ts, shared by MobileScheduleView and ServiceSessionModal (the two maps were identical).
