+++ src/lib/schedule/constants.ts
@@ -1,0 +1,15 @@
+/**
+ * Duration choices (value → label) offered when creating or resizing a service
+ * session. Shared by the desktop `ServiceSessionModal` and the mobile add/edit
+ * sheets so the option list stays in sync in one place.
+ */
+export const SERVICE_DURATION_OPTIONS = new Map([
+  ['5', '5 minutes'],
+  ['10', '10 minutes'],
+  ['15', '15 minutes'],
+  ['20', '20 minutes'],
+  ['30', '30 minutes'],
+  ['45', '45 minutes'],
+  ['60', '60 minutes'],
+  ['90', '90 minutes'],
+])
